requirements.txt.txt
streamlit
phi
python-dotenv
